tests: restructure utils in Jest and Cypress
