+++ __tests__/boundary.test.tsx
@@ -5,7 +5,7 @@ import ReactThreeTestRenderer from '@react-three/test-renderer';
 import React from 'react';
 import * as THREE from 'three';
 
-import { ChildrenType, TestProps, checkBoxMesh, makeSnakeHeadProps } from '@/testutils/snake';
+import { ChildrenType, TestProps, checkBoxMesh, makeSnakeHeadProps } from './utils/snake';
 import { Head } from '@/app/components/snake/Head';
 import { BoundaryProps, makeBoundaryProps, Boundary } from '@/app/components/snake/Boundary';
 import { gameOrigin, halfBoundary, isBoundaryHit } from '@/app/utils/snake';
